Move annuaire routes to /annuaire, and split resources
refs #1288
