+++ app/controllers/annuaire/institutions_controller.rb
@@ -1,0 +1,17 @@
+module  Annuaire
+  class InstitutionsController < ApplicationController
+    def index
+      authorize Institution, :index?
+
+      @institutions = Institution.all
+        .order(:slug)
+        .preload([institutions_subjects: :theme], :not_deleted_antennes, :not_deleted_advisors)
+
+      @wide_layout = true
+    end
+
+    def show
+      redirect_to institution_subjects_path(params[:slug])
+    end
+  end
+end
