+++ app/controllers/annuaire/subjects_controller.rb
@@ -1,0 +1,6 @@
+module Annuaire
+  class SubjectsController < BaseController
+    def index
+    end
+  end
+end
