+++ app/policies/institution_policy.rb
@@ -19,19 +19,11 @@ class InstitutionPolicy < ApplicationPolicy
     admin?
   end
 
-  def import_antennes?
+  def import?
     admin?
   end
 
-  def import_advisors?
-    admin?
-  end
-
-  def import_antennes_create?
-    admin?
-  end
-
-  def import_advisors_create?
+  def import_create?
     admin?
   end
 end
