+++ app/views/annuaire/advisors/_table.html.haml
