+++ app/views/annuaire/antennes/_table.html.haml
