+++ app/views/annuaire/antennes/index.html.haml
@@ -2,6 +2,6 @@
 - content_for :header, render('header', institution: @institution)
 - content_for :menu, render('menu', institution: @institution)
 
-= render 'advisors_import_buttons'
+= render 'import_buttons'
 
-= render 'advisors_table', institutions_subjects: @institutions_subjects, advisors: @advisors
+= render 'table', antennes: @antennes
