+++ app/views/annuaire/base/_menu.html.haml
@@ -1,0 +1,3 @@
+= active_link_to Institution.human_attribute_name(:subjects), institution_subjects_path(institution), { class: 'item' }
+= active_link_to Institution.human_attribute_name(:antennes), institution_antennes_path(institution), { class: 'item' }
+= active_link_to Institution.human_attribute_name(:advisors), institution_advisors_path(institution), { class: 'item' }
