+++ app/views/annuaire/institutions/index.html.haml
@@ -2,4 +2,4 @@
 - meta title: title
 %h1.ui.header= title
 
-= render 'institutions_table', institutions: @institutions
+= render 'table', institutions: @institutions
