+++ app/views/application/_navbar.html.haml
@@ -50,7 +50,7 @@
                 = t('.admin')
               = active_link_to institutions_path, class: 'item' do
                 %i.landmark.icon
-                = t('.institutions')
+                = t('.annuaire')
               = active_link_to inserts_partner_tools_path, class: 'item' do
                 %i.wrench.icon
                 = t('.partner_tools')
