--- app/views/institutions/_advisors_import_buttons.html.haml
@@ -1,9 +1,0 @@
-.ui.secondary.fitted.menu
-  .item
-    = link_to advisors_institution_path(format: 'csv'), class: 'ui labeled icon' do
-      %i.download.icon
-      = t('download')
-  .item
-    = link_to import_advisors_institution_path, class: 'ui labeled icon' do
-      %i.upload.icon
-      = t('import')
