--- app/views/institutions/_menu.html.haml
@@ -1,3 +1,0 @@
-= active_link_to Institution.human_attribute_name(:subjects), subjects_institution_path(institution), { class: 'item' }
-= active_link_to Institution.human_attribute_name(:antennes), antennes_institution_path(institution), { class: 'item' }
-= active_link_to Institution.human_attribute_name(:advisors), advisors_institution_path(institution), { class: 'item' }
